+++ CHANGELOG.md
@@ -450,3 +450,4 @@
 * Change `Substructured` constraint type according to new `Substructure` definition
 * Define `:::` - type operator for almost anything higher kinded
 * Redefine `Zipper` - add possible moves to type definition
+* Define experimental `Fastenable` type family to have it as `Constraint`